wip: Add index to codes and use for both edit and create
This is an exploration to create a smoother transition between the create
and edit routes. We are currently using the same page component (codes/index)
to handle both the create and edit pages - staying on the same route. This
approach works, in terms of being able to save and update the QR code, but
falls short of navigating to meaningful, linkable URLS.